backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1

# Database
psycopg2-binary>=2.9.9

# AI/ML
google-generativeai>=0.3.2

# HTTP
requests>=2.31.0
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0

# WebSockets (for family helper)
channels>=4.0.0
channels-redis>=4.1.0

# Utilities
python-dateutil>=2.8.2
